se añade archivo que calcula promedios de peso e imc de cada paciente ingresado

+++ paciente_arreglo.cpp
@@ -7,7 +7,7 @@ using namespace std;
 struct Paciente {
     string nombre;
     int edad;
-    double altura;
+    int altura;
     double peso;
 };
